fix(studio): compile Windows 0.2.6 release and skip icon SVG lint
studio-v0.2.5 Windows CI failed: SHChangeNotify wants i32, and
show_window is still unstable on the release rustc. Agent stays on
WslLaunch. Biome no longer a11y-lints ICO source SVGs.

+++ apps/studio/src-tauri/src/presentment.rs
@@ -141,7 +141,7 @@ fn notify_shell_icon_changed() {
         };
         unsafe {
             SHChangeNotify(
-                SHCNE_ASSOCCHANGED,
+                SHCNE_ASSOCCHANGED as i32,
                 SHCNF_IDLIST | SHCNF_FLUSH,
                 std::ptr::null(),
                 std::ptr::null(),
